implement sqlx-cli (sqlx migrate)
this is a standard with sqlx so might as well use it. could help with
production deploys and also having multiple dev environments.
move db/schema.sql to a new migration file. the file required an
alteration so that it would find the "_sqlx_migrations" file.
remove script/schema.sh as it is obsolete.
rename the PG_URL environment variable to DATABASE_URL to automatically
be recognized by sqlx-cli.