implement custom errors for Credentials
i used a variety of online resources to design this.
create "pub enum Error" and give it the elements UsernameFormatError,
PasswordLengthError, and PasswordIncludesUsernameError. then return
these as Err() and handle them in the calling function from main.rs.
this is a skill i want to learn eventually, and it allows more
granularity on input validations when compared to a simple bool.