improved WebSocket message structure
instead of passing a JSON string around interally, create a new
PostMessage struct which is constructed using a Post and the html
partial that is rendered of it. do authorization analysis on this
struct (especially whether or not a user owns it, for rejected posts)
and only convert it to json upon sending to the JavaScript.

instead of "postSubmitted" and other actions, simply send the post
status which is all that is needed.

add a ?1 to the script.js inclusion so that it busts cache.

clean up the anon_uuid macro code so that we don't need "mut jar" in
actions where it is not pertinent.

rename some functions in javascript.

add is_admin and is_author helpers, since we use this rather
complicated logic multiple times.

add "autocomplete" attribute to some text inputs because chrome asks
for it, although we use "current-password" for our "hybrid" password
form.