restructure code with new routes module
I think too much was in main.rs.
I made a lot of things in main.rs "pub" so that they could easily
be imported with "use crate::*".
This also gives a tidier separation between public and private
functions in both files.
All constants, "mod" statements, "use" statements are going to the top
of the file in a consistent manner.