rename User to Account
create a new User struct which includes both Account and anon_uuid.
this allows some additional simplicity in our route handlers and also
our model files as typically Account and anon info are coupled.

a user is any user of the site; a user may be anon, and a user may
optionally have an account. this is tighter terminology which is more
accurate, which is important for comprehension.

i think the benefit is worth the added complexity of having a "meta"
struct.

rename the "users" table to "accounts"; rename all relevant indexes
and constraints, and the "user_id" column on posts to "account_id".

add a temporary migration for moving the "user" cookie to "account".
in time we will remove this.