allow setting of a PORT env var
of 7878 is not to someone's taste. I got 7878 from the example web
server provided in the Rust book. other good reasons to change the
port might be running multiple versions of the app.

update README a little regarding this setting and also web server
config (web sockets and IP forwarding).