check pathname in JS
without this, it was setting the title to "undefined" due to our title
restoration system.

it was also initializing websockets unnecessarily on the login and
registration pages.

i thought of setting an "id" attr on the body tag, but realized I could
just use conditionals by pathname which is more intuitive.