don't raise error if send fails
if receiver_count() is zero then calling send() will actually cause an
error which results in a 502 for the user. This has caused this to
happen consistently when no users are connected via the websocket,
which happens if a user has Javascript disabled. This fix will make
the site work properly for non-Javascript users. calling ok() will
convert the Result into an Option which will be ignored harmlessly.