make init module in main.rs
I think this is slightly cleaner because the methods in this are
otherwise included in router.rs because of the crate::* usage.

I don't like that crate::main() is used in router.rs but I think this
is better than how it was before.

The "use" statements, constants, and module definitions in main.rs are
"global" variables for use in router.rs. Having an "init" module makes
sense for its relevant functions which are related to the app booting.
AppState is also needed in router.rs.

This also allows us to "use" the trace layer structs in the init
module, which is much cleaner than including them in router.rs.