remove streaming, write file the easy way
Not only is streaming overly complicated, we have decided it will be safer
to encrypt the file in memory prior to writing it to disk.

Files will only be unencrypted on-the-fly for logged-in admins, or they will
be stored in unencrypted and public format after admin approval.

Add a 10MB limit for uploads via DefaultBodyLimit layer.