encrypt files using cocoon
https://docs.rs/cocoon/latest/cocoon/index.html

This library seems pretty simple and uses AES-256 encryption.

I notice it has a performance hit upon encrypting files, but that is probably
acceptable. If it is slow to decrypt, that might be a problem.

I renamed "IP_SALT" in the .env (and .env.example) to SECRET_KEY, since I want
to reuse this variable for general encryption purposes.

The check for SECRET_KEY length now occurs at boot rather than when hashing an
IP.