remove owner from initial schema commit
otherwise it prevents having any other postgres username besides the one we
initially began this project as.

I will need to update the checksums in my production databases so that sqlx
doesn't complain. this is a hack that is only appropriate for alpha software,
but it is a necessary one as there is no other solution i can find.

we discovered this problem when setting up a second instance of the app.