restore some custom scrollbar styling
Use only the standard "scrollbar-color" declaration.
Place color styles (and scrollbar-color) on the html element so that it affects
the browser scrollbar.
This does affect textarea (automatically), but it does not affect select boxes.
I was unable to find a way to style select boxes and keep the buttons.
This preserves the normal width and buttons so as to be less annoying.
Based on a revert of commit 105a1e73b5b2346ba8a45d210fbc40e833a01ec7.