Dramatically refactor websocket and templates
It's strange these are connected, but it's because of the template variable
re-configuration.

I wanted to use "user" in templates, and this made me realize I was rendering
the "post" partial in the websocket incorrectly. I don't need to call it twice
with admin on and off, I just need to render it for the specific user in the
receiver.

I was then able to get rid of PostMessage entirely and just send Post.

I do a proper init_user on the login and registration forms now.

Also rename user.jinja to profile.jinja for clarity.