apply phone margins everywhere
also decrease maximum flex size to 400px.
This allows four images per row on 1080p.
It's best because the maximum width of a thumbnail is 800px, which on retina
screen requires 1600px. Twice the flex size is the maximum width (on list
page). With the new margin system it looks better than only two "boutique"
images, and with the new solo system with luxurious display, there is less
need for such display on the list page.