replace session token with user token
This is a permanent cookie instead of a session cookie.
The concept of the "session" cookie is increasingly archaic (although I still
allow it as an option for account login). The user token represents a specific
browser, more accurately.
We were facing a serious problem in the Brave browser where when it booted up
(under its default configuration, no less) it would have cleared the session
cookie, but still have the old session token in its cached HTML. Firefox and
Chrome, which have this feature as optional, simply do not clear session
cookies when it is enabled.
Brave is increasingly a popular browser so we need to support it.
The "reload" Javascript solution is a hack. Yes it provides interim data, but
this is not that important. The CSRF token is important.
A separate CSRF token generated on each request would be even better, but do
we need it?