Revert "replace session token with user token"
This reverts commit 47081654ac356aa8cd68438410511d4eee37c2aa.

This was a mistake.
I needed to set the Cache-Control header to "no-store" for the app in nginx.
I tried this before but did it wrong.
This solves all of the problems I was experiencing related to caching, and is
standard behavior on webapps. I feel a little foolish. But at least it is
fixed.