make solo_post handler and template
the "index" action was getting unwieldy, and nothing about the solo page is
really an "index" action.

index will still handle page posts, however.

Differentiate more clearly between the two in the CSS by applying a class to
div#posts even during the index action.

And display a message at the top of the screen when a page is loaded.