center media
I just think it looks nicer this way.
Add a "media" div surrounding media, and a "youtube" div surrounding embeds.
Use "fit-content" where appropriate.
Keep the -8px margin specifically on the media tags so that the default margin
remains elsewhere.
Use div.post specificity for that margin rather than div#posts, or else it is
not easy for the div.youtube a img to override it without !important.
Restore a more sensible 90vh max height on images.