center everything
the body tag itself is centered now, and the default body margin is applied
instead to the html tag. yes, this is legal, although likely frowned upon.

a new "content" div is added to every post to facilitate centering its entire
contents correctly within its "post" container. This way the body, media, and
arrow will all be aligned correctly. Unfortunately this means adding an extra
tab to the relevant partials, unless I can find a way later to avoid it.