remove data-fetch attribute
unnecessary because we add fetch to all forms on the index page, and otherwise
ignore this attribute.
if in the future we need to avoid adding fetch for some reason, it would
probably be better to add a data-no-fetch attribute, or specify directly in
the JS which forms to add fetch to.
data-confirm by contrast I think still has value.