use /post/{key} for solo page
it was always kind of hacky using /{key}, to the point it diminishes from the
aesthetic benefit of it.

Especially because /settings and /register are both technically valid post
keys, even at eight characters each!

The problem became clear in a branch when we wanted to match the solo page
route in JS, and found we could not do it without matching other pages.