set min-width on images
this fixes the safari rendering issue, which was caused by it not knowing the
image widths. It would render the width based on the timestamp on the solo page
(I believe), and then insert the image, adding an unexpected margin. This
margin would go away on refresh or certain other page actions, but ruined the
first load of a solo page.

Setting a min-width of 300px, which is desired anyway, avoids this issue.