add byte_slice custom filter
using the substring [:limit] in minijinja goes by a character index, not a byte
index (like the rust slice does). thus we need a special function just to give
us the rust byte slicing ability in jinja.

long-term, this should be removed, and the intro should be cached as a separate
database column.