add lib.rs and migrate.rs
this is an attempt to create two separate binaries, apabbs and migrate, which
use shared code. the shared code, which is essentially the entire application,
is in lib.rs. the main.rs file will now just referenced the public elements of
the apabbs library, and so will the migrate.rs file.