add special migration function type
I mostly used copilot magic to figure this out. It's because the function
pointers for the various migration functions actually have distinct types.
We need to use some janky hackwork to make it work properly.

Also note I am passing in a raw PgPool instead of a reference, and especially
not a mutable reference to PgConnection.

This commit is WIP and contains some changes related to the thumbnail
downloading migration.