hack _sqlx_migrations table
this will remove an old migration that was already executed. immediately after
this is deployed, we will delete the migration file to prevent it ever running
again.

this only works because we are applying it solely to our single-dev dev
environments and our production server. this situation should be avoided
entirely going forwards.

when this change is applied to our alternate historical data which has not yet
run the deleted migration, it will simply bypass it as if it never existed.

we are still in alpha, but improved migration support is a step towards getting
in beta.