add migration at an earlier timestamp
this is a hack but less abrasive than the others. it will do nothing on our
production data, but is necessary for migrating our prior data set. this is
because in our production data, youtube thumbnail urls were migrated at the
same time as the old bash script for downloading youtube thumbnails was
executed. this essentially replaces that bash script's database functionality
in the canonical sqlx migration timeline.

it replaces the system of updating the youtube thumbnail image urls via the
download_youtube_thumbnails rust migration as that was not necessary.