use Rust-based alphanumeric key generator
delete the old SQL function which was used for this purpose, as it is overly
complicated and we prefer to avoid making custom PL/pgSQL functions now.

also, ensure that the generated key does not already exist. although we have
hundreds of trillions of key possibilities with only eight alphanumeric chars,
it is still better to be safe than sorry.