apply interim post checking to non-approved posts
otherwise i end up with a situation where a pending post does not get
reflected as approved since i put it to sleep.

for simplicity in query selecting, i added the "approved" class back to the
post divs.

i did not manually test this much but feel confident that it should work.