use _opt more consistently
even in jinja and javascript.

also add a security check when it comes to rejected posts. should not have been
a problem before (as a user would not have received a rejected post by another
user to display), but better safe than sorry.

i almost thought of getting rid of _opt, but the fact is I still think it is
the more responsible behavior.