stricter use of opt in jinja and scripts
Don't print out _opt variables in jinja. you have to "unwrap" them by
assigning them to a variable without the opt suffix. this basically mimics
the rust behavior in an unofficial way. it does make the templates a little
more complex on the surface, but I think it increases logical consistency.
It allows for some shorter lines as well.