use H.264 for compatibility
Firefox does not support H.265, and Safari does not support AV-1.
I am using the "medium" preset for now. We can switch to "fast" if it feels
too slow in production.
Add a timestamp of 0.001 seconds to videos so that their first frame shows up
as a thumbnail in Safari.
Update template display a little for video thumbnail support.