discord embed wip
i'd like it to support both images and videos.
we will likely have to add a "poster" image before we can continue this.

this commit makes a significant change to .env, renaming SITE_NAME to HOST.
this is more accurate and makes more sense when we have to provide the host
in HTML urls. this will of course require updating production's .env file
before deployment.