fix chrome bug after post update
the "poster" attribute was being ignored improperly after chrome update.

it seems the poster attribute is (according to spec) supposed to be a URL.
we however, are using a path which is customary everywhere else, e.g. for img
sources.

Firefox and Safari handle this just fine, but Chrome requires I re-set the
poster attribute, and it ends up setting the poster attribute as a full URL
because it comes from a DOM attribute.

I think this is good enough because it works.