replace non_mac_webkit() with analyze_user_agent()
Create a UserAgent struct to store two booleans, whether the User-Agent header
shows the user is on a Mac and whether they are on a Chromium-based browser.

Include this in an Option, as User-Agent may not be specified in the request
headers.

I moved away from saying "WebKit" because we are specifically looking for
Chromium browsers more precisely.

This is a prerequisite to showing compatibility versions of videos in
non-Chromium browsers.