run screenshot generator as an hourly cron job
this avoids most of the problems related to generating a screenshot after every
relevant home page update. There is not sufficient reason to be so precise with
the screenshout, hourly is good enough and is much easier to manage.

This should be paired later with a change to the index.jinja template that adds
a query string (representing the last hour) to the screenshot.webp so that it
is cache-busted in Discord or elsewhere.