disable --no-sandbox option
I'm not sure what it does, just seems to be less secure. Used to be required,
but seems to work for me locally without it.

This is part of me trying to debug these dbus related errors in production. The
errors don't prevent the screenshot from being made, but they do clutter up the
error log and I'm not sure why.