more screenshot tweaks
disable stderr output entirely. I found no way to suppress the DBus errors
short of actually creating a DBus daemon, and that is not worth it.

print a success message in stdout.

restore --disable-gpu, as one of the stderr warnings related to GPU.

clone output_path_str so that it can be used both in the blocking task and also
in the success message.