remove all spawn_blocking for new processes
It turns out this was unnecessary, as creating a new process already uses a new
CPU core, and tokio handles waiting for the process to finish using standard
async tools.

I also removed all Send + Sync, except the two cases it is necessary, and not
because of spawn_blocking but rather because I pass a Result over an await
and access it later.