move transaction and websocket utilities to lib.rs
I want these functions to be available in our "model" code; they are
technically more associated with AppState than with Axum request handling.

This is part of a restructuring that is necessary for our rerun_failed_tasks
function in migrate.rs.

(Also rename continue_failed_processing to rerun_failed_tasks)