standardize expect statements
I want them to match the intended format (a "should" statement) and also look
good if they show up as an error. The messages are generic and describe error
type more than code context, which is fine as these will come with a stack
trace and line numbers.

Use unwrap() instead of expect() for Options.