Improve WebSocket reconnection
Before it was doing 30-second intervals, which is too long for wake-from-sleep
or a server restart due to deployment. Instead start at 5 seconds and double
the duration up to a maximum of 60 seconds.

I increased it to 30 seconds in the first place because Firefox has an
automatic system for reducing request frequency (not found in Chrome). This
system however resulted in multiple requests being pending at a time, so when
the connection was restored, it would connect multiple times. This was fixed
by ensuring the webSocket readyState is not open or pending in initWebSocket().

I was also able to remove the hacky "webSocketOpen" variable, instead checking
for the existence of reconnectTimeout.

Lastly, as cleanup, I adjusted the handler functions for webSocket to be
defined via "onopen" (and similar) instead of addEventListener, as we do not
expect there to ever be multiple handlers for these events.