Tweak CSS
Rename main.index class to "list", as it is used on pages besides index, and
is more conventionally descriptive.

Move media brightness adjustments in dark mode to the media section, and apply
them only to div.media content. This means it will not affect for example,
YouTube logos, which are small and not bothersome, or any other future images
we add to the layout.

Use "main.list" to describe list page content instead of "div#posts", which is
less clear in its intention.

Add some comments where helpful.