add vipsheader_field
this will allow us to extract any header field from vipsheader, although we only
use it for i32s currently. thus I have to parse it twice, which is perhaps
worse? ah well, i decided this was cleaner. this also adds an additional heap
allocation for the String. basically, this is a downgrade!