username approval proof of concept
had to rename the "action" input to "intent" to avoid conflicting with the form
action in scripts.

add some messaging to the settings page for pending accounts.

add scripting to handle removal of account review list items after successful
fetch.

add 404 support to scripting, as it is possible it may occur.

use div.row within li, as li has its own display type and we need the flex
display type on the row.