prevent creating a new account when one is pending
this also changes the nature of the flooding autoban code, which now only
checks for pending posts. As it is impossible to create more than one pending
account per IP address at a time, there is no point to checking for account
registration flooding.

Move some constants related to test IP addresses around in the tests as well,
and add a couple new ones to avoid conflicts. Remove the test_ip_hash() function
which was not necessary.