only run addPostHidingButtons on root
this should only appear when a user who does NOT own a post is dynamically
updated a post which was previously displayed but is now rejected, banned or
reported. that can only happen on the homepage via websocket.

The "hide post" button should show for dynamically reported posts for every
user that is not an admin, not just mods.

Also remove some outdated "helper" variables from post.jinja.